chore: speed up ide iframe by spliting code
