+++ src/parser.js
@@ -7,6 +7,7 @@ var compilers = require("./compiler/compilers");
 var { typecheck, printSignature } = require("./typecheck");
 var { expandMacros, extractMacros } = require("./macro.js");
 var { defaultImportReader } = require("./reader");
+var { evalCompiled } = require("./execute");
 
 const defaultTrustedHosts = [
   "https://raw.githubusercontent.com/LingDong-/wenyan-lang/master"
@@ -809,61 +810,6 @@ function compile(arg1, arg2, arg3) {
   }
 
   return targ;
-}
-
-function isLangSupportedForEval(lang) {
-  if (lang !== "js")
-    throw new Error(
-      `Executing for target language "${lang}" is not supported in current environment`
-    );
-  return true;
-}
-
-function hanzinize(value) {
-  if (typeof value == "number") {
-    return num2hanzi(value);
-  } else if (typeof value == "boolean") {
-    return bool2hanzi(value);
-  } else if (Array.isArray(value)) {
-    return value.map(i => hanzinize(i)).join("。");
-  } else {
-    return value;
-  }
-}
-
-function outputHanziWrapper(log, outputHanzi) {
-  return function output(...args) {
-    log(...args.map(i => (outputHanzi ? hanzinize(i) : i)));
-  };
-}
-
-function evalCompiled(compiledCode, options = {}) {
-  const {
-    outputHanzi = true,
-    scoped = false,
-    lang = "js",
-    output = console.log
-  } = options;
-
-  isLangSupportedForEval(lang);
-
-  let code = compiledCode;
-
-  (() => {
-    const _console_log = console.log;
-    console.log = outputHanziWrapper(output, outputHanzi);
-    try {
-      if (!scoped && "window" in this) {
-        window.eval(code);
-      } else {
-        eval(code);
-      }
-    } catch (e) {
-      throw e;
-    } finally {
-      console.log = _console_log;
-    }
-  })();
 }
 
 function execute(source, options = {}) {
